Removed function exists check
This was failing under certain situations and causing the raw short
code to be displayed. I’ve prefixed the function and named it so that
it’s fairly unlikely to run into conflicts. You might want to wrap the
whole plugin in a class and break up this function a bit.